backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.7
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
httpx>=0.25.0
cryptography>=41.0.0